fixed some setters in model (forgot to swap from in to Long)

+++ src/main/java/it/univaq/we/internshipTutor/model/Internship.java
@@ -83,7 +83,7 @@ public class Internship {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
