+++ src/main/java/it/univaq/we/internshipTutor/model/Professor.java
@@ -41,7 +41,7 @@ public class Professor {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
